pnnx save ncnn bin with fp16 storage (#3715)

+++ tools/pnnx/tests/ncnn/test_nn_Conv3d.py
@@ -46,7 +46,7 @@ class Model(nn.Module):
         return x
 
 def test():
-    net = Model()
+    net = Model().half().float()
     net.eval()
 
     torch.manual_seed(0)
